+++ Editor/TextureCompressor/Core/Models/TextureInfo.cs
@@ -14,6 +14,7 @@ namespace dev.limitex.avatar.compressor.editor.texture
         FrozenSkip,
         RuntimeGenerated,
         ExcludedPath,
+        UnknownUncompressedProperty,
     }
 
     /// <summary>